Extract branch protection into a separate module
