+++ bin/update_branch_protection.rb
@@ -7,16 +7,13 @@ require 'manageiq/release'
 require 'trollop'
 
 opts = Trollop.options do
-  opt :branch, "The branch to protect.", :type => :string
+  opt :branch, "The branch to protect.", :type => :string, :required => true
   opt :repo, "The repo to update. If not passed, will try all repos for the branch specified.", :type => :string
 
   opt :dry_run, "", :default => false
 end
-Trollop.die("Must pass either --repo or --branch") unless opts[:branch_given] || opts[:repo_given]
 
 ManageIQ::Release.each_repo(opts[:repo], opts[:branch]) do |repo|
-  ManageIQ::Release::UpdateRepoSettings.new(repo.github_repo, opts.slice(:dry_run)).run
-
-  skip = !opts[:branch_given] || (opts[:branch] != "master" && repo.options[:has_real_releases])
+  skip = opts[:branch] != "master" && repo.options[:has_real_releases]
   ManageIQ::Release::UpdateBranchProtection.new(repo.github_repo, opts.slice(:branch, :dry_run)).run unless skip
 end
