+++ lib/manageiq/release.rb
@@ -11,6 +11,7 @@ require 'manageiq/release/release_branch'
 require 'manageiq/release/release_tag'
 require 'manageiq/release/rename_labels'
 require 'manageiq/release/rename_sprint_milestones'
+require 'manageiq/release/update_branch_protection'
 require 'manageiq/release/update_labels'
 require 'manageiq/release/update_repo_settings'
 require 'manageiq/release/update_sprint_milestones'
